pyproject.toml
[tool.poetry]
name = "quartermaster"
version = "0.1.1"
description = "USB "
authors = ["Tim Laurence <timdaman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.1"
huey = "^2.1.3"
requests = "^2.22.0"
django-bootstrap4 = "^1.1.1"
redis = "^3.3.11"
djangorestframework = "^3.11.0"
psycopg2-binary = "^2.8.4"
paramiko = "^2.7.1" # TODO: Move this to driver
filelock = "^3.0.12"
coverage = "^5.0.1"


[tool.poetry.dev-dependencies]
mypy = "^0.761"
mypy-extensions = "0.4.3"
pytest = "^5.3.2"
pytest-django = "^3.7.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
